+++ GBShop/GBShop/App/AppDelegate.swift
@@ -7,14 +7,37 @@
 
 import UIKit
 import Alamofire
+import Swinject
+
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-  
-  let requestFactory = RequestFactory()
+  let container: Container = {
+    let container = Container()
+    container.register(AbstractErrorParser.self) { _ in ErrorParser() }
+    container.register(Auth.self) { r in
+      lazy var commonSession: Session = {
+        let configuration = URLSessionConfiguration.default
+        configuration.httpShouldSetCookies = false
+        configuration.headers = .default
+        let manager = Session(configuration: configuration)
+        return manager
+      }()
+      let sessionQueue = DispatchQueue.global(qos: .utility)
+
+      return Auth(errorParser: r.resolve(
+        AbstractErrorParser.self)!,
+        sessionManager: commonSession,
+        queue: sessionQueue
+      )
+    }
+    
+    return container
+  }()
   
   func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-    let auth = requestFactory.makeAuthRequestFatory()
+    let auth = container.resolve(Auth.self)!
+
     auth.login(userName: "Somebody", password: "mypassword", completionHandler: responseResult)
     auth.logout(userId: 123, completionHandler: responseResult)
     auth.registerUser(
