--- GBShop/GBShop/App/BusinessLogic/Network/RequestFactory.swift
@@ -1,32 +1,0 @@
-//
-//  RequestFactory.swift
-//  GBShop
-//
-//  Created by Valera Vvedenskiy on 17.02.2022.
-//
-
-import Foundation
-import Alamofire
-
-class RequestFactory {
-  
-  func makeErrorParser() -> AbstractErrorParser {
-    return ErrorParser()
-  }
-  
-  lazy var commonSession: Session = {
-    let configuration = URLSessionConfiguration.default
-    configuration.httpShouldSetCookies = false
-    configuration.headers = .default
-    let manager = Session(configuration: configuration)
-    return manager
-  }()
-  
-  let sessionQueue = DispatchQueue.global(qos: .utility)
-  
-  func makeAuthRequestFatory() -> AuthRequestFactory {
-    let errorParser = makeErrorParser()
-    return Auth(errorParser: errorParser, sessionManager: commonSession, queue: sessionQueue)
-  }
-}
-
