+++ GBShop/GBShop/App/Flows/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  GBShop
+//
+//  Created by Valera Vvedenskiy on 15.02.2022.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    // Do any additional setup after loading the view.
+  }
+
+
+}
+
